+++ src/Utility/PageHelper.php
@@ -2,6 +2,7 @@
 
 namespace Foodsharing\Utility;
 
+use Detection\MobileDetect;
 use Foodsharing\Lib\Session;
 use Foodsharing\Modules\Core\DBConstants\Foodsaver\Role;
 use Foodsharing\Modules\Core\DBConstants\Region\RegionIDs;
@@ -17,11 +18,11 @@ use Foodsharing\Permissions\RegionPermissions;
 use Foodsharing\Permissions\ReportPermissions;
 use Foodsharing\Permissions\StorePermissions;
 use Foodsharing\Permissions\WorkGroupPermissions;
-use JetBrains\PhpStorm\Deprecated;
 use Twig\Environment;
 
 final class PageHelper
 {
+    private const SESSION_MOBIL_REPRESENTATION_FIELD_NAME = 'mob';
     private string $add_css = '';
     private string $content_main = '';
     private string $content_right = '';
@@ -60,6 +61,16 @@ final class PageHelper
         private readonly ProfilePermissions $profilePermissions,
         private readonly RegionGateway $regionGateway
     ) {
+    }
+
+    public function isMob(): bool
+    {
+        if (!$this->session->has(self::SESSION_MOBIL_REPRESENTATION_FIELD_NAME)) {
+            $mobdet = new MobileDetect();
+            $this->session->set(self::SESSION_MOBIL_REPRESENTATION_FIELD_NAME, $mobdet->isMobile());
+        }
+
+        return $this->session->get(self::SESSION_MOBIL_REPRESENTATION_FIELD_NAME);
     }
 
     public function generateAndGetGlobalViewData(): array
@@ -105,7 +116,7 @@ final class PageHelper
             'route' => $page,
             'dev' => FS_ENV == 'dev',
             'hidden' => $this->hidden,
-            'isMob' => $this->session->isMob(),
+            'isMob' => $this->isMob(),
             'footer' => $this->getFooter(),
             'HTTP_HOST' => $_SERVER['HTTP_HOST'] ?? BASE_URL,
             'content' => [